requirements-dev.txt
flake8==6.1.0
pytest==7.4.2
pytest-django==4.5.2
